Self-publish registration information once a cluster leader is available.
This basically means the Consul takes care of itself, instead of realying
on Terraform pushing information in the KV store *after* deployment. Much
more self-contained and ultimately simple.

For this, I brought over consul-get-or-set from nubis-storage (hint: move in nubis-base)

Also, feels like we are duplicating a lot of logic left and right to wait
for Consul to be up and running, should probably find a generic solution to
handle this in base

Fixes #66