Generate the puppet-foo needed to manage the Platform ACL / Anonymous ACLs
Create the ACLs, and since we are blacklisting for now, we don't need
to be very specific, the Platform token gets *.* since we allow by default,
and it's the anonymous one that gets denied explicitely. Will need inverting
at some point, once things are neat and tidy, so we can instead just deny by
default and all.

Fixes #139